Fix /W3 msvc warnings and cmake linter

+++ examples/gui/ofdm/render_ofdm_demod.cpp
@@ -252,8 +252,8 @@ void RenderSynchronisation(OFDM_Demod& demod) {
             ImPlot::PlotLine("Impulse response", buf.data(), (int)buf.size());
 
             // Plot useful markers for coarse freq sync using freq correlation
-            const int coarse_freq_offset = (int)std::round(demod.GetCoarseFrequencyOffset() * Fs);
-            const int max_coarse_freq_offset = cfg.sync.max_coarse_freq_correction_norm * Fs;
+            const int coarse_freq_offset = int(std::round(demod.GetCoarseFrequencyOffset() * Fs));
+            const int max_coarse_freq_offset = int(cfg.sync.max_coarse_freq_correction_norm * Fs);
             const int freq_fft_bin = int(Fs / float(params.nb_fft));
             const int peak_offset_x = -coarse_freq_offset / freq_fft_bin;
             const int max_offset_x = max_coarse_freq_offset / freq_fft_bin;
